pyproject.toml
[build-system]
requires = ['setuptools>=80.9.0', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'tortoise-embeddings'
version = '0.0.1'
description = 'Adds pgvector support to TortoiseORM. '
readme = 'README.md'
requires-python = '>=3.14'
authors = [
    {name = 'Kaoticz', email = '1812311+Kaoticz@users.noreply.github.com'},
]
dependencies = [
    'tortoise-orm==0.24.2',
    'aerich==0.9.2',
    'asyncpg==0.31.0',
    'pgvector==0.4.2',
    'numpy==2.4.4',
]

[tool.setuptools.packages.find]
include = ['tortoise_embeddings', 'tortoise_embeddings.*']

[tool.mypy]
python_version = '3.14'
files = ['tortoise_embeddings/']
strict = true
local_partial_types = true
exclude = '.venv/'
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pyright]
typeCheckingMode = 'standard'
analyzeUnannotatedFunctions = true
reportDeprecated = true
reportUnannotatedClassAttribute = true
reportImplicitOverride = true
reportUnknownLambdaType = false

[tool.pytest.ini_options]
testpaths = ['tests']
addopts = ['-v', '--showlocals']
asyncio_mode = 'auto'
filterwarnings = [
    'ignore:_UnionGenericAlias is deprecated:DeprecationWarning:google.genai.types',
]

[project.optional-dependencies]
dev = [
    'setuptools',
    'build',
    'mypy==1.20.0',
    'basedpyright==1.39.0',
    'pytest==9.0.2',
    'pytest-asyncio==1.3.0',
    'google-genai==1.70.0',
    'python-dotenv==1.2.2'
]
